chore(dashboard): share target filters
